POS embeddings, word embeddings, corpus helper
Now using the PTB POS tag set; I had to scrape the list of part-of-speech tags manually from the NLTK website.

Word embeddings as well as POS embeddings have been trained and saved. The word embeddings are the result of training a single LSTM on 5-word sequences from Penn Treebank using a 10k word vocabulary. After 30k training rounds with a fixed learning rate of 1e-4, the 100-round average perplexity reached the low 200's (not great, but not awful considering it was only trained on 5-word sequences).

The POS embeddings were literally trained just using an autoencoder; I briefly considered not creating the embedding at all, because the POS tags are so low-dimension already... but decided to create the embedding anyway.

The embedding matrices have wildly different dimensions, which is OK. The full language model will simple concatenate the results of the separate matrix multiplies between the word embedding and POS embedding, and pass that on to the rec